suggestions from review (edit functions, code duplication,typos,etc)
