+++ projects/python/perception/pose_estimation/high_resolution_pose_estimation/README.md
