+++ projects/python/perception/pose_estimation/high_resolution_pose_estimation/demos/benchmarking_demo.py
@@ -53,11 +53,11 @@ if __name__ == '__main__':
                                                          mobilenet_use_stride=stride, half_precision=half_precision,
                                                          first_pass_height=base_height1,
                                                          second_pass_height=base_height2)
-    pose_estimator.download(path=".", verbose=True)
+    pose_estimator.download(path="", verbose=True)
     pose_estimator.load("openpose_default")
 
     # Download one sample image
-    pose_estimator.download(path=".", mode="test_data")
+    pose_estimator.download(path="", mode="test_data")
     image_path = join("temp", "dataset", "image", "000000000785_1080.jpg")
     img = cv2.imread(image_path)
