+++ projects/python/perception/pose_estimation/lightweight_open_pose/demos/benchmarking_demo.py
@@ -13,11 +13,12 @@
 # limitations under the License.
 
 import cv2
+import time
 from opendr.perception.pose_estimation import LightweightOpenPoseLearner
-from opendr.perception.pose_estimation import draw
-from opendr.engine.data import Image
 import argparse
 from os.path import join
+from tqdm import tqdm
+import numpy as np
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
@@ -39,20 +40,34 @@ if __name__ == '__main__':
 
     pose_estimator = LightweightOpenPoseLearner(device=device, num_refinement_stages=stages,
                                                 mobilenet_use_stride=stride, half_precision=half_precision)
-    pose_estimator.download(path=".", verbose=True)
+    pose_estimator.download(path="", verbose=True)
     pose_estimator.load("openpose_default")
 
     # Download one sample image
-    pose_estimator.download(path=".", mode="test_data")
+    pose_estimator.download(path="", mode="test_data")
     image_path = join("temp", "dataset", "image", "000000000785.jpg")
-    img = Image.open(image_path)
+    img = cv2.imread(image_path)
 
     if onnx:
         pose_estimator.optimize()
 
-    poses = pose_estimator.infer(img)
-    img_cv = img.opencv()
-    for pose in poses:
-        draw(img_cv, pose)
-    cv2.imshow('Results', img_cv)
-    cv2.waitKey(0)
+    fps_list = []
+    print("Benchmarking...")
+    for i in tqdm(range(50)):
+        start_time = time.perf_counter()
+        # Perform inference
+        poses = pose_estimator.infer(img)
+        end_time = time.perf_counter()
+        fps_list.append(1.0 / (end_time - start_time))
+    print("Average FPS: %.2f" % (np.mean(fps_list)))
+
+    # If pynvml is available, try to get memory stats for cuda
+    try:
+        if 'cuda' in device:
+            from pynvml import nvmlInit, nvmlDeviceGetMemoryInfo, nvmlDeviceGetHandleByIndex
+
+            nvmlInit()
+            info = nvmlDeviceGetMemoryInfo(nvmlDeviceGetHandleByIndex(0))
+            print("Memory allocated: %.2f MB " % (info.used / 1024 ** 2))
+    except ImportError:
+        pass
