+++ projects/python/perception/pose_estimation/lightweight_open_pose/demos/webcam_demo.py
@@ -1,0 +1,101 @@
+# Copyright 2020-2022 OpenDR European Project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import cv2
+import time
+from opendr.perception.pose_estimation import LightweightOpenPoseLearner
+from opendr.perception.pose_estimation import draw
+import argparse
+
+
+class VideoReader(object):
+    def __init__(self, file_name):
+        self.file_name = file_name
+        try:  # OpenCV needs int to read from webcam
+            self.file_name = int(file_name)
+        except ValueError:
+            pass
+
+    def __iter__(self):
+        self.cap = cv2.VideoCapture(self.file_name)
+        if not self.cap.isOpened():
+            raise IOError('Video {} cannot be opened'.format(self.file_name))
+        return self
+
+    def __next__(self):
+        was_read, img = self.cap.read()
+        if not was_read:
+            raise StopIteration
+        return img
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--onnx", help="Use ONNX", default=False, action="store_true")
+    parser.add_argument("--device", help="Device to use (cpu, cuda)", type=str, default="cuda")
+    parser.add_argument("--accelerate", help="Enables acceleration flags (e.g., stride)", default=False,
+                        action="store_true")
+    args = parser.parse_args()
+
+    onnx, device = args.onnx, args.device
+    accelerate = args.accelerate
+
+    onnx, device, accelerate = args.onnx, args.device, args.accelerate
+    if accelerate:
+        stride = True
+        stages = 0
+        half_precision = True
+    else:
+        stride = False
+        stages = 2
+        half_precision = False
+
+    pose_estimator = LightweightOpenPoseLearner(device=device, num_refinement_stages=stages,
+                                                mobilenet_use_stride=stride, half_precision=half_precision)
+    pose_estimator.download(path="", verbose=True)
+    pose_estimator.load("openpose_default")
+
+    if onnx:
+        pose_estimator.optimize()
+
+    # Use the first camera available on the system
+    image_provider = VideoReader(0)
+
+    try:
+        counter, avg_fps = 0, 0
+        for img in image_provider:
+
+            start_time = time.perf_counter()
+
+            # Perform inference
+            poses = pose_estimator.infer(img)
+            end_time = time.perf_counter()
+            fps = 1.0 / (end_time - start_time)
+
+            for pose in poses:
+                draw(img, pose)
+
+            # Calculate a running average on FPS
+            avg_fps = 0.8 * fps + 0.2 * fps
+
+            # Wait a few frames for FPS to stabilize
+            if counter > 5:
+                image = cv2.putText(img, "FPS: %.2f" % (avg_fps,), (50, 50), cv2.FONT_HERSHEY_SIMPLEX,
+                                    1, (255, 0, 0), 2, cv2.LINE_AA)
+
+            cv2.imshow('Result', img)
+            cv2.waitKey(1)
+            counter += 1
+    except:
+        print("Average inference fps: ", avg_fps)
