+++ projects/python/perception/pose_estimation/lightweight_open_pose/jetbot/evaluate.sh
