+++ projects/python/perception/pose_estimation/lightweight_open_pose/jetbot/simulation_pose/worlds/textures/eyelashes04.png
